notify_hitokuchi/requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
idna==2.8
lxml==4.3.4
py==1.8.0
python-dotenv==0.10.3
requests==2.22.0
retry==0.9.2
soupsieve==1.9.2
urllib3==1.25.3
